+++ docs/changelog.rst
@@ -21,7 +21,7 @@ Here we list what's new in ``pyprep``.
 
 .. _changes_0_6_0:
 
-Version 0.6.0 (unreleased)
+Version 0.6.0 (2026-02-26)
 --------------------------
 
 Changelog